feat(Checkbox): Replace deprecated implementation
