+++ packages/react-component-library/cypress/selectors/Checkbox.ts
