+++ packages/react-component-library/cypress/selectors/DismissibleBanner.ts
@@ -1,5 +1,5 @@
 export default {
-  checkbox: '[data-testid="dimissiblebanner-dontshow"]',
+  checkbox: '[data-testid="checkbox-input"]',
   description: '[data-testid="dimissiblebanner-description"]',
   dismiss: '[data-testid="dimissiblebanner-dismiss"]',
   title: '[data-testid="dimissiblebanner-title"]',
