+++ packages/react-component-library/src/components/Checkbox/Checkbox.stories.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 
-import { Checkbox } from '.'
+import { Checkbox, CheckboxProps } from '.'
+import { CHECKBOX_VARIANT } from './constants'
 
 export default {
   component: Checkbox,
@@ -11,35 +12,83 @@ export default {
   },
 } as ComponentMeta<typeof Checkbox>
 
-export const Default: ComponentStory<typeof Checkbox> = (props) => (
+const Template: ComponentStory<typeof Checkbox> = (props) => (
   <Checkbox {...props} />
 )
 
+const MultipleItemsTemplate: ComponentStory<typeof Checkbox> = (props) => {
+  function getProps(i: number): CheckboxProps {
+    return {
+      ...props,
+      label: `${props.label} ${i}`,
+      name: `${props.name}-${i}`,
+    }
+  }
+
+  return (
+    <>
+      <Checkbox {...getProps(1)} />
+      <Checkbox {...getProps(2)} />
+      <Checkbox {...getProps(3)} />
+    </>
+  )
+}
+
+export const Default = Template.bind({})
 Default.args = {
   id: undefined,
   label: 'Default checkbox',
   name: 'default',
-  isChecked: true,
 }
 
-export const Disabled: ComponentStory<typeof Checkbox> = (props) => (
-  <Checkbox {...props} />
-)
+export const Checked = Template.bind({})
+Checked.args = {
+  id: undefined,
+  defaultChecked: true,
+  label: 'Checked',
+  name: 'checked',
+}
 
-Disabled.args = {
+export const DisabledUnchecked = Template.bind({})
+DisabledUnchecked.storyName = 'Disabled, unchecked'
+DisabledUnchecked.args = {
   id: undefined,
   isDisabled: true,
-  label: 'Disabled checkbox',
+  label: 'Disabled, unchecked',
   name: 'disabled',
 }
 
-export const Invalid: ComponentStory<typeof Checkbox> = (props) => (
-  <Checkbox {...props} />
-)
+export const DisabledChecked = Template.bind({})
+DisabledChecked.storyName = 'Disabled, checked'
+DisabledChecked.args = {
+  id: undefined,
+  defaultChecked: true,
+  isDisabled: true,
+  label: 'Disabled, checked',
+  name: 'disabled',
+}
 
+export const Invalid = Template.bind({})
 Invalid.args = {
   id: undefined,
   label: 'Invalid checkbox',
   name: 'invalid',
   isInvalid: true,
+}
+
+export const NoContainer = MultipleItemsTemplate.bind({})
+NoContainer.args = {
+  id: undefined,
+  label: 'Item without container',
+  name: 'no-container',
+  variant: CHECKBOX_VARIANT.NO_CONTAINER,
+}
+
+export const WithDescription = Template.bind({})
+WithDescription.args = {
+  id: undefined,
+  description:
+    'She must have hidden the plans in the escape pod. Send a detachment down to retrieve them.',
+  label: 'With description',
+  name: 'with-description',
 }
