+++ packages/react-component-library/src/components/Checkbox/Checkbox.test.tsx
@@ -1,8 +1,11 @@
 import React from 'react'
 import '@testing-library/jest-dom/extend-expect'
+import 'jest-styled-components'
+import { ColorNeutral200 } from '@defencedigital/design-tokens'
 import { render, RenderResult } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
 
-import { Checkbox } from '.'
+import { Checkbox, CHECKBOX_VARIANT } from '.'
 import { FieldProps } from '../../common/FieldProps'
 import { FormProps } from '../../common/FormProps'
 
@@ -11,14 +14,16 @@ describe('Checkbox', () => {
   let form: FormProps
   let label: string
   let checkbox: RenderResult
+  let onChangeSpy: jest.Mock
 
   beforeEach(() => {
     label = ''
+    onChangeSpy = jest.fn()
 
     field = {
       name: 'example1',
       value: 'false',
-      onChange: jest.fn(),
+      onChange: onChangeSpy,
       onBlur: jest.fn(),
     }
 
@@ -29,6 +34,8 @@ describe('Checkbox', () => {
   })
 
   describe('when a field has no errors, a label and a value', () => {
+    let input: HTMLElement
+
     beforeEach(() => {
       label = 'My Label 1'
       field.value = 'false'
@@ -41,21 +48,76 @@ describe('Checkbox', () => {
           onChange={field.onChange}
         />
       )
+
+      input = checkbox.getByTestId('checkbox-input')
     })
 
     it('should render a field with a label', () => {
-      expect(checkbox.getByTestId('label')).toHaveTextContent('My Label 1')
-    })
-
-    it('should populate the field value', () => {
-      expect(checkbox.queryByTestId('checkbox')).toHaveAttribute(
-        'value',
-        'false'
+      expect(checkbox.getByTestId('checkbox-label')).toHaveTextContent(
+        'My Label 1'
       )
     })
 
+    it('should not render a description', () => {
+      expect(checkbox.queryAllByTestId('checkbox-description')).toHaveLength(0)
+    })
+
+    it('should populate the field value', () => {
+      expect(input).toHaveAttribute('value', 'false')
+    })
+
     it('should not initially render as checked', () => {
-      expect(checkbox.getByTestId('checkbox')).not.toHaveAttribute('checked')
+      expect(input).not.toBeChecked()
+    })
+
+    describe('and tab is pressed', () => {
+      beforeEach(() => {
+        userEvent.tab()
+      })
+
+      it('focuses the input', () => {
+        expect(input).toHaveFocus()
+      })
+
+      describe('and space is pressed', () => {
+        beforeEach(() => {
+          userEvent.keyboard('[Space]')
+        })
+
+        it('checks the input', () => {
+          expect(input).toBeChecked()
+        })
+
+        it('calls onChange once', () => {
+          expect(field.onChange).toHaveBeenCalledTimes(1)
+        })
+
+        describe('and space is pressed again', () => {
+          beforeEach(() => {
+            onChangeSpy.mockReset()
+            userEvent.keyboard('[Space]')
+          })
+
+          it('unchecks the input', () => {
+            expect(input).not.toBeChecked()
+          })
+
+          it('calls onChange once', () => {
+            expect(field.onChange).toHaveBeenCalledTimes(1)
+          })
+        })
+      })
+    })
+
+    it('should render a container', () => {
+      expect(checkbox.getByTestId('checkbox')).toHaveStyleRule(
+        'border',
+        `1px solid ${ColorNeutral200}`
+      )
+      expect(checkbox.getByTestId('checkbox')).toHaveStyleRule(
+        'border-radius',
+        '15px'
+      )
     })
   })
 
@@ -76,7 +138,7 @@ describe('Checkbox', () => {
     })
 
     it('should initially render as checked', () => {
-      expect(checkbox.getByTestId('checkbox')).toHaveAttribute('checked')
+      expect(checkbox.getByTestId('checkbox-input')).toBeChecked()
     })
   })
 
@@ -97,7 +159,7 @@ describe('Checkbox', () => {
     })
 
     it('should initially render as checked', () => {
-      expect(checkbox.getByTestId('checkbox')).toHaveAttribute('checked')
+      expect(checkbox.getByTestId('checkbox-input')).toBeChecked()
     })
   })
 
@@ -115,7 +177,7 @@ describe('Checkbox', () => {
     })
 
     it('should attach the class to the wrapper', () => {
-      expect(checkbox.queryByTestId('container')).toHaveClass('test')
+      expect(checkbox.queryByTestId('checkbox')).toHaveClass('test')
     })
   })
 
@@ -133,11 +195,17 @@ describe('Checkbox', () => {
     })
 
     it('should attach the id to the field', () => {
-      expect(checkbox.queryByTestId('checkbox')).toHaveAttribute('id', 'test')
+      expect(checkbox.queryByTestId('checkbox-input')).toHaveAttribute(
+        'id',
+        'test'
+      )
     })
 
     it('should associate the label to the field with the custom id', () => {
-      expect(checkbox.queryByTestId('label')).toHaveAttribute('for', 'test')
+      expect(checkbox.queryByTestId('checkbox-label')).toHaveAttribute(
+        'for',
+        'test'
+      )
     })
   })
 
@@ -155,9 +223,62 @@ describe('Checkbox', () => {
     })
 
     it('should drill the arbitrary prop', () => {
-      expect(checkbox.queryByTestId('checkbox')).toHaveAttribute(
+      expect(checkbox.queryByTestId('checkbox-input')).toHaveAttribute(
         'data-arbitrary',
         'arbitrary'
+      )
+    })
+  })
+
+  describe('when a field has a description', () => {
+    beforeEach(() => {
+      checkbox = render(
+        <Checkbox description="Description" label="Label" name={field.name} />
+      )
+    })
+
+    it('should render a description', () => {
+      expect(checkbox.getByTestId('checkbox-description')).toHaveTextContent(
+        'Description'
+      )
+    })
+  })
+
+  describe('when a field has a description that is arbitrary JSX', () => {
+    beforeEach(() => {
+      checkbox = render(
+        <Checkbox
+          description={<div>Arbitrary content</div>}
+          label="Label"
+          name={field.name}
+        />
+      )
+    })
+
+    it('should render a arbitrary description', () => {
+      expect(checkbox.getByText('Arbitrary content')).toBeInTheDocument()
+    })
+  })
+
+  describe('when a field does not have a container', () => {
+    beforeEach(() => {
+      checkbox = render(
+        <Checkbox
+          label="Label"
+          name={field.name}
+          variant={CHECKBOX_VARIANT.NO_CONTAINER}
+        />
+      )
+    })
+
+    it('should not render a container', () => {
+      expect(checkbox.getByTestId('checkbox')).not.toHaveStyleRule(
+        'border',
+        `1px solid ${ColorNeutral200}`
+      )
+      expect(checkbox.getByTestId('checkbox')).not.toHaveStyleRule(
+        'border-radius',
+        '15px'
       )
     })
   })
