+++ packages/react-component-library/src/components/Checkbox/Checkbox.tsx
@@ -1,13 +1,21 @@
-import React, { forwardRef } from 'react'
+import React, { useState, useRef } from 'react'
 import { v4 as uuidv4 } from 'uuid'
+import mergeRefs from 'react-merge-refs'
 
+import { CHECKBOX_VARIANT } from './constants'
 import { ComponentWithClass } from '../../common/ComponentWithClass'
 import { InputValidationProps } from '../../common/InputValidationProps'
 import { StyledCheckbox } from './partials/StyledCheckbox'
-import { StyledOuterWrapper } from './partials/StyledOuterWrapper'
-import { StyledLabel } from './partials/StyledLabel'
-import { StyledInput } from './partials/StyledInput'
 import { StyledCheckmark } from './partials/StyledCheckmark'
+import { StyledCheckboxWrapper } from './partials/StyledCheckboxWrapper'
+import { StyledDescription } from './partials/StyledDescription'
+import { StyledInput } from './partials/StyledInput'
+import { StyledLabel } from './partials/StyledLabel'
+import { StyledOuterWrapper } from './partials/StyledOuterWrapper'
+
+type CheckboxVariantType =
+  | typeof CHECKBOX_VARIANT.DEFAULT
+  | typeof CHECKBOX_VARIANT.NO_CONTAINER
 
 export interface CheckboxProps
   extends ComponentWithClass,
@@ -25,10 +33,9 @@ export interface CheckboxProps
    */
   defaultChecked?: boolean
   /**
-   * Toggles whether or note the component is marked as checked.
-   * @deprecated
+   * Optional description to display below the label.
    */
-  isChecked?: boolean
+  description?: React.ReactNode
   /**
    * Toggles whether the component is disabled or not (preventing user interaction).
    */
@@ -53,56 +60,98 @@ export interface CheckboxProps
    * Optional HTML `value` attribute associated with the component.
    */
   value?: string
+  /**
+   * Optional variant to set container visibility.
+   */
+  variant?: CheckboxVariantType
 }
 
-export const Checkbox = forwardRef<HTMLInputElement, CheckboxProps>(
+export const Checkbox = React.forwardRef<HTMLInputElement, CheckboxProps>(
   (
     {
       className = '',
-      id = uuidv4(),
       checked,
       defaultChecked,
-      isChecked,
+      description,
+      id = uuidv4(),
       isDisabled,
+      isInvalid,
       label,
       name,
       onBlur,
       onChange,
       value,
-      isInvalid,
+      variant = CHECKBOX_VARIANT.DEFAULT,
       ...rest
     },
     ref
   ) => {
+    const localRef = useRef<HTMLInputElement>(null)
+    const [isChecked, setIsChecked] = useState(defaultChecked || checked)
+
+    const handleClick = (event: React.MouseEvent<HTMLDivElement>) => {
+      if (event.target !== localRef.current) {
+        localRef.current?.click()
+      }
+    }
+
+    const handleKeyUp = (_: React.KeyboardEvent) => {
+      localRef.current?.focus()
+    }
+
+    const handleOnChange = (e: React.FormEvent<HTMLInputElement>) => {
+      setIsChecked(!isChecked)
+
+      if (onChange) {
+        onChange(e)
+      }
+    }
+
+    const hasContainer = variant !== CHECKBOX_VARIANT.NO_CONTAINER
+
     return (
-      <StyledCheckbox
-        $isDisabled={isDisabled}
-        $isInvalid={isInvalid}
-        className={className}
-        data-testid="container"
-      >
-        <StyledOuterWrapper>
-          <StyledLabel htmlFor={id} data-testid="label">
-            <StyledInput
-              $isDisabled={isDisabled}
-              ref={ref}
-              id={id}
-              type="checkbox"
-              name={name}
-              value={value}
-              defaultChecked={isChecked || defaultChecked}
-              onChange={onChange}
-              onBlur={onBlur}
-              disabled={isDisabled}
-              checked={checked}
-              data-testid="checkbox"
-              {...rest}
-            />
-            <StyledCheckmark />
-            {label}
-          </StyledLabel>
-        </StyledOuterWrapper>
-      </StyledCheckbox>
+      <StyledCheckboxWrapper>
+        <StyledCheckbox
+          $hasContainer={hasContainer}
+          $isDisabled={isDisabled}
+          $isInvalid={isInvalid}
+          $isChecked={isChecked}
+          className={className}
+          onClick={handleClick}
+          onKeyUp={handleKeyUp}
+          data-testid="checkbox"
+        >
+          <StyledOuterWrapper $hasContainer={hasContainer}>
+            <StyledLabel
+              $hasContainer={hasContainer}
+              htmlFor={id}
+              data-testid="checkbox-label"
+            >
+              <StyledInput
+                ref={mergeRefs([localRef, ref])}
+                id={id}
+                type="checkbox"
+                name={name}
+                value={value}
+                defaultChecked={defaultChecked}
+                onChange={handleOnChange}
+                onBlur={onBlur}
+                disabled={isDisabled}
+                checked={checked}
+                {...rest}
+                data-testid="checkbox-input"
+              />
+              <StyledCheckmark $hasContainer={hasContainer} />
+              {label}
+              {description && (
+                <StyledDescription data-testid="checkbox-description">
+                  {description}
+                </StyledDescription>
+              )}
+            </StyledLabel>
+          </StyledOuterWrapper>
+        </StyledCheckbox>
+      </StyledCheckboxWrapper>
     )
   }
 )
