+++ packages/react-component-library/src/components/Checkbox/index.ts
@@ -1,1 +1,2 @@
 export * from './Checkbox'
+export * from './constants'
