+++ packages/react-component-library/src/components/Checkbox/partials/StyledCheckbox.tsx
@@ -1,37 +1,71 @@
-import React from 'react'
 import { selectors } from '@defencedigital/design-tokens'
 import styled, { css } from 'styled-components'
 
-import { StyledLabel } from './StyledLabel'
-
-const { color, fontSize, spacing } = selectors
+const { color, fontSize } = selectors
 
 export interface StyledCheckboxProps {
+  $hasContainer?: boolean
   $isDisabled?: boolean
   $isInvalid?: boolean
+  $isChecked?: boolean
 }
 
 export const StyledCheckbox = styled.div<StyledCheckboxProps>`
-  display: block;
+  display: inline-flex;
   position: relative;
-  padding-left: ${spacing('11')};
   font-size: ${fontSize('base')};
   user-select: none;
+  cursor: pointer;
 
-  ${({ $isDisabled }) =>
-    $isDisabled &&
-    css`
-      ${StyledLabel} {
-        cursor: not-allowed;
-        color: ${color('neutral', '200')};
-      }
-    `}
+  * {
+    cursor: pointer;
+  }
 
   ${({ $isInvalid }) =>
     $isInvalid &&
     css`
-      ${StyledLabel} {
-        color: ${color('danger', '700')};
+      border-color: ${color('danger', '800')};
+      box-shadow: 0 0 0 2px ${color('danger', '800')};
+    `}
+
+  ${({ $hasContainer, $isChecked }) =>
+    $hasContainer &&
+    $isChecked &&
+    css`
+      background-color: ${color('action', '000')};
+    `}
+
+  ${({ $hasContainer }) =>
+    $hasContainer &&
+    css`
+      border: 1px solid ${color('neutral', '200')};
+      border-radius: 15px;
+
+      &:focus-within,
+      &:active {
+        outline: none;
+        border-color: ${color('action', '500')};
+        box-shadow: 0 0 0 2px ${color('action', '500')},
+          0 0 0 5px ${color('action', '100')};
+      }
+    `}
+
+  ${({ $isDisabled }) =>
+    $isDisabled &&
+    css`
+      cursor: not-allowed;
+
+      * {
+        cursor: not-allowed;
+      }
+
+      background-color: ${color('neutral', '000')};
+      border-color: transparent;
+
+      &:focus,
+      &:active {
+        border-color: transparent;
+        box-shadow: none;
       }
     `}
 `
