+++ packages/react-component-library/src/components/Checkbox/partials/StyledDescription.tsx
