--- packages/react-component-library/src/components/CheckboxE/index.ts
@@ -1,2 +1,0 @@
-export * from './CheckboxE'
-export * from './constants'
