--- packages/react-component-library/src/components/CheckboxE/partials/StyledOuterWrapper.tsx
@@ -1,22 +1,0 @@
-import styled, { css } from 'styled-components'
-
-interface StyledOuterWrapperProps {
-  $hasContainer?: boolean
-}
-
-export const StyledOuterWrapper = styled.div<StyledOuterWrapperProps>`
-  display: inline-flex;
-  align-items: center;
-  flex-direction: row;
-  border-radius: 15px;
-
-  &:active {
-    pointer-events: none;
-  }
-
-  ${({ $hasContainer }) =>
-    $hasContainer &&
-    css`
-      min-height: 44px;
-    `}
-`
