--- packages/react-component-library/src/components/CheckboxEnhanced/index.ts
@@ -1,1 +1,0 @@
-export * from './CheckboxEnhanced'
