+++ packages/react-component-library/src/components/DismissibleBanner/partials/StyledDontShow.tsx
@@ -1,11 +1,5 @@
-import { selectors } from '@defencedigital/design-tokens'
 import styled from 'styled-components'
 
 import { Checkbox } from '../../Checkbox'
 
-const { spacing } = selectors
-
-export const StyledDontShow = styled(Checkbox)`
-  padding-top: ${spacing('3')};
-  padding-bottom: ${spacing('3')};
-`
+export const StyledDontShow = styled(Checkbox)``
