+++ packages/react-component-library/src/components/Fieldset/partials/StyledFieldset.tsx
@@ -1,8 +1,8 @@
 import styled, { css } from 'styled-components'
 import { selectors } from '@defencedigital/design-tokens'
 
-import { StyledCheckboxWrapper } from '../../CheckboxE/partials/StyledCheckboxWrapper'
-import { StyledCheckbox } from '../../CheckboxE/partials/StyledCheckbox'
+import { StyledCheckboxWrapper } from '../../Checkbox/partials/StyledCheckboxWrapper'
+import { StyledCheckbox } from '../../Checkbox/partials/StyledCheckbox'
 import { StyledRadioWrapper } from '../../RadioE/partials/StyledRadioWrapper'
 import { StyledRadio } from '../../RadioE/partials/StyledRadio'
 import { StyledField } from '../../Field/partials/StyledField'
