+++ packages/react-component-library/src/index.ts
@@ -17,8 +17,6 @@ export * from './components/Button'
 export * from './components/ButtonGroup'
 export * from './components/CardFrame'
 export * from './components/Checkbox'
-export * from './components/CheckboxE'
-export * from './components/CheckboxEnhanced'
 export * from './components/ContextMenu'
 export * from './components/DatePicker'
 export * from './components/DatePickerE'
